feat: split chain_spec
